requirements.txt
mcp
mitmproxy
yt-dlp
openai
python-dotenv
starlette
uvicorn